epwa2709it6/line03/requirements.txt
line-bot-sdk
flask
Pillow
pytesseract
Werkzeug
